requirements.txt
certifi==2018.8.24
chardet==3.0.4
future==0.16.0
idna==2.7
python-telegram-bot==10.1.0
requests==2.19.1
urllib3==1.23
